Updated project dependencies and plugin library versions
The lein-compojure template is very simple to get started with, although it
seems the libraries are a little behind the latest.  The project runs
successfully without upgrading versions.  It is usually better to use the latest
stable versions of these libraries to pick up any fixes.

The latest stable versions were found via https://clojars.org/.

Consider submitting a pull request to update the lein-compojure template project
on Github.