Created svg-components namespace
Create the status-monitor.svg-components namespace to contain a whole range of
SVG elements that could be used for the dashboard.  Initially just a demo image
has been created.  The demo image has been added to the status-monitor dashboard
to show how well SVG components are rendered in a web page.