Added missing path of CIDER logo, changed viewbox size
The left hand side of the glass in the CIDER logo was missing as I had not
included the path from the HTML code.

The path needs a little adjustment as its not translating correctly from the
original HTML.  Or it could be the HTML I am using is not correct either.

Increased the size of the viewbox as the image was far too large on the page.